consensus: Rename agreement module to quorum module
- Rename step.rs (quorum module) to task.rs
- Replace agreement with quorum label
