+++ consensus/src/quorum.rs
@@ -4,5 +4,5 @@
 //
 // Copyright (c) DUSK NETWORK. All rights reserved.
 
-pub(crate) mod step;
+pub(crate) mod task;
 pub mod verifiers;
